Chat2Anything-admin/requirements.txt
Flask
Flask-APScheduler
Flask-Migrate
Flask-Login
Flask_Reuploaded
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
psutil
Flask-Mail
Pillow
cryptography